refactor(phase-1): introduce src/core/ foundation with compat shims
Creates the zero-domain-logic foundation every extension will eventually
import from. Old module paths continue to work for one release via thin
re-export shims, so no extension code changes in this commit.

Added
- src/core/__init__.py — package marker + module map.
- src/core/logging.py  — CustomFormatter + init_logger/get_logger
                         (moved from src/logutil.py).
- src/core/db.py       — MongoManager singleton + mongo_manager
                         (moved from src/mongodb.py).
- src/core/config.py   — load_config/save_config/save_module_field +
                         new ConfigStore (singleton with get()/save_full()/
                         reload()/subscribe()). Legacy save paths now
                         fire subscriber notifications so extensions can
                         observe WebUI edits without a bot restart.
                         Atomic-write logic from Phase 0 carried over.
- src/core/errors.py   — shared exception base classes: BotError,
                         ConfigError, DatabaseError, ValidationError,
                         IntegrationError, HttpError. Existing feature
                         exceptions will migrate onto this tree in later PRs.
- src/core/http.py     — HttpClient singleton around a single long-lived
                         aiohttp.ClientSession, plus fetch() with the same
                         signature as src.utils.fetch. Raises HttpError on
                         final failure (still an Exception subclass so
                         existing except-Exception sites keep working).

Shimmed
- src/logutil.py       — re-exports src.core.logging.
- src/mongodb.py       — re-exports src.core.db.
- src/config_manager.py — re-exports src.core.config.
- src/utils.py         — fetch() now re-exported from src.core.http; the
                         inline implementation is deleted. asyncio import
                         dropped (no longer used).

Verified
- python -m py_compile across src/, extensions/, main.py (raiderio.py uses
  Py3.12-only f-string syntax and is skipped locally — unchanged).
- Import probe confirms every legacy ``from src.X import Y`` still resolves
  and ConfigStore + HttpError wire up cleanly.
